add routes to admin & added auth to posts and admin

+++ routes/admin.js
@@ -2,17 +2,34 @@ const express = require('express');
 const router = express.Router();
 const { User } = require('../models');
 
-/* Admin DELETS USER */
 
+/* ADMIN GET ALL USERS - displays all users */
+router.get('/', async (req, res, next) => {
+    const user = req.user;
+        if (!user){
+            res.status(403).send('Please Log In!');
+            return;
+        };
+
+        if (user.user_name != 'ad'){
+            res.status(403).send('Access denied!');
+            return;
+        };
+
+        User.findAll()
+        .then( allUsers => {
+            res.json(allUsers)
+        }).catch( err => {
+            res.status(500).send(err)
+        })
+});
+
+/* ADMIN DELETES USER */
 router.delete('/:id', async (req, res, next) => {
-    const userId = pareseInt(req.params.id)
+    const userId = parseInt(req.params.id)
 
         if (!userId || userId <= 0){
             res.status(400).send('Please insert an id');
-            return;
-        };
-        if (!User.admin){
-            res.status(403).send('Access denied!');
             return;
         };
         const user = req.user;
@@ -20,7 +37,12 @@ router.delete('/:id', async (req, res, next) => {
                 res.status(403).send('Please log in!');
                 return;
             };
-
+//          Admin Check
+        if (user.user_name != 'ad'){
+            res.status(403).send('Access denied!');
+            return;
+        };
+        
         User.destroy({
             where: {
                 id: userId
@@ -32,12 +54,5 @@ router.delete('/:id', async (req, res, next) => {
         })
 });
 
-
-
-
-/* ADMIN GET ALL USERS - displays all users */
-router.get('/', async (req, res, next) => {
-
-});
 
 module.exports = router;
